requirements.txt
flask
pyodbc
azure-storage-blob
